Upport of #4762
